+++ eras/babbage/impl/CHANGELOG.md
@@ -24,6 +24,9 @@
 ### `testlib`
 
 * Add `Test.Cardano.Ledger.Babbage.Imp.PoolSpec` module with `babbageEraSpecificSpec`
+* In `Test.Cardano.Ledger.Babbage.Examples`:
+  - Remove `mkBabbageBasedExampleTx`, `exampleBabbageBasedTxBody`
+  - Add `exampleBabbageBasedTx`, `exampleBabbageTx`, `exampleBabbageToConwayTx`
 
 ## 1.13.0.0